Make it so you can see black in undrawn areas
I'm not sure whether CameraClearFlags.Depth is faster than the default
or not... seems like it should be because the screen doesn't have to
fill with a color first. But for now I'll take it out so you can see
what's going on.

Performance oriented users could switch this to
CameraClearFlags.Nothing, which might perform faster.