Removing quads when there are too many empty ones
This was a pain to implement because I bumped into a weird Unity issue
where a gameObject will get stuck to the bottom of the draw order if you
Clear its mesh and give it less vertices than it used to have... or
something like that. To be honest I'm not sure what happened, but
setting someMaterial.renderQueue to do the render order works much
better, and makes more sense.