Fixing text vertical alignment
This was a pain because some fonts handle vertical spacing differently,
so I'm not sure this will work for everyone, but it should work for most
people.

There's also a shouldVerticallyAlignUsingBase property I added to the
FTextParams for fonts that need to be adjusted to be vertically
centered.

This solves issue #23