Load atlas images from byte data
This is an alternative to using normal Unity image assets for textures
(which still works fine). The reason this is a problem is because on iOS
Unity either stores images as compressed and lossy pvrtc files, or
uncompressed and super-huge raw bitmap files. This makes for larger
on-device app sizes and longer build times.

This solution simply stores png byte data in a file, and then loads it
into a texture at runtime. I haven't tested to see if the memory hit is
larger, but so far it seems to work great.