Changing shader names and adding banana particles
I changed the Futile shader names so they actually exist in the Futile
folder, and as part of that I added the unity particles to the Banana
Demo Project.

I still haven't tested the particles on mobile, but they work great on
PC, the only issue on mobile will be that I'm creating a new particle
emitter every time rather than reusing the old one, so it could
definitely be more efficient.