Fixing threading issue with RXProfiler
This fixes #147

I was using a normal Timer, which actually runs on another thread, so
crazy things had the potential to happen when iterating through an array
while the timer was firing and removing items from the array in a
different thread.

I've changed it to just use Unity's Update() loop instead, so that
there's no more threading.