Adding FRepeatMeshSprite
This is a mesh-based version of FRepeatSprite and is quite useful... it
basically allows you to do tiling textures but using elements from
within an atlas. Like FRepeatSprite it also has .scrollX and .scrollY
properties.

I'm not 100% sure I'll keep this name, I may merge it into FRepeatSprite
and use the ideal implementation based on whether the atlas element is a
single image or not.